Avoid a few warnings on the alpha

+++ usr.sbin/ppp/auth.c
@@ -17,7 +17,7 @@
  * IMPLIED WARRANTIES, INCLUDING, WITHOUT LIMITATION, THE IMPLIED
  * WARRANTIES OF MERCHANTIBILITY AND FITNESS FOR A PARTICULAR PURPOSE.
  *
- * $Id: auth.c,v 1.41 1999/02/20 01:12:45 brian Exp $
+ * $Id: auth.c,v 1.42 1999/02/26 21:28:06 brian Exp $
  *
  *	TODO:
  *		o Implement check against with registered IP addresses.
@@ -335,7 +335,7 @@ auth_ReadHeader(struct authinfo *authp, struct mbuf *bp)
   } else {
     authp->in.hdr.length = htons(0);
     log_Printf(LogWARN, "auth_ReadHeader: Short packet header (%d > %d) !\n",
-               sizeof authp->in.hdr, len);
+               (int)(sizeof authp->in.hdr), len);
   }
 
   mbuf_Free(bp);
